requirements.txt
django
djangorestframework
djoser
pendulum
ipython
ipdb